requirements.txt
pyTelegramBotAPI==4.12.0
Flask==2.3.2
ffmpeg-python==0.2.0
Pillow==9.5.0
gunicorn==20.1.0